// src/app/shared/models/persona.ts
export class Persona {
    id?: string;
    ApellidoMaterno: string;
    ApellidoPaterno: string;
    CentroPoblado: string;
    CodigoHogar: string;
    DNI: string;
    Direccion: string;
    Distrito: string;
    Edad: string;
    EsTitular: boolean;
    FechaEntrevista: string;
    FechaNacimiento: string;
    FrecuenciaAtencion: string;
    Genero: string;
    IngresosEconomicos: string;
    NivelInstruccion: string;
    NombreEntrevistador: string;
    Nombres: string;
    Ocupacion: string;
    ProgramaSocial: string;
    ServiciosMedicos: string;
    SituacionEconomica: string;
    Telefono: string;
    VinculoFamiliar: string;
    idEntrevistador: string;
    idFamilia: string;
    Estado: boolean;
}